Add support for solidity-coverage

+++ src/setup-ganache.ts
@@ -2,7 +2,7 @@ import path from 'path';
 import { fork, ChildProcess } from 'child_process';
 
 import { accountsConfig } from './accounts';
-import config, { Config } from './config';
+import config from './config';
 
 interface ErrorMessage {
   type: 'error';
@@ -31,42 +31,55 @@ export type NodeOptions = {
 };
 
 export default async function (): Promise<string> {
-  const server = fork(path.join(__dirname, 'ganache-server'), [], {
-    // Prevent the child process from also being started in inspect mode, which
-    // would cause issues due to parent and child sharing the port.
-    // See https://github.com/OpenZeppelin/openzeppelin-test-environment/pull/23
-    execArgv: process.execArgv.filter((opt) => opt !== '--inspect'),
-  });
+  if (!config.coverage) {
+    const server = fork(path.join(__dirname, 'ganache-server'), [], {
+      // Prevent the child process from also being started in inspect mode, which
+      // would cause issues due to parent and child sharing the port.
+      // See https://github.com/OpenZeppelin/openzeppelin-test-environment/pull/23
+      execArgv: process.execArgv.filter((opt) => opt !== '--inspect'),
+    });
 
-  const gasPrice = typeof config.node.gasPrice === 'string' ? config.node.gasPrice : undefined;
+    const gasPrice = typeof config.node.gasPrice === 'string' ? config.node.gasPrice : undefined;
 
-  const options: NodeOptions = {
-    ...config.node,
-    gasPrice,
-    accounts: accountsConfig,
-    coverage: config.coverage,
-  };
-  server.send(options);
+    const options: NodeOptions = {
+      ...config.node,
+      gasPrice,
+      accounts: accountsConfig,
+      coverage: config.coverage,
+    };
+    server.send(options);
 
-  const messageReceived: Promise<Message> = new Promise(
-    (resolve): ChildProcess => {
-      return server.once('message', resolve);
-    },
-  );
+    const messageReceived: Promise<Message> = new Promise(
+      (resolve): ChildProcess => {
+        return server.once('message', resolve);
+      },
+    );
 
-  const message = await messageReceived;
+    const message = await messageReceived;
 
-  switch (message.type) {
-    case 'ready':
-      if (server.channel) {
-        server.channel.unref();
-      }
-      server.unref();
-      return `http://localhost:${message.port}`;
-    case 'error':
-      server.kill();
-      throw new Error('Unhandled server error');
-    default:
-      throw new Error(`Uknown server message: '${message}'`);
+    switch (message.type) {
+      case 'ready':
+        if (server.channel) {
+          server.channel.unref();
+        }
+        server.unref();
+        return `http://localhost:${message.port}`;
+      case 'error':
+        server.kill();
+        throw new Error('Unhandled server error');
+      default:
+        throw new Error(`Uknown server message: '${message}'`);
+    }
+  } else {
+    if (process.send === undefined) {
+      throw new Error('Module must be started through child_process.fork for solidity-coverage support.');
+    }
+    process.send(accountsConfig);
+    const address: string = await new Promise(
+      (resolve): NodeJS.Process => {
+        return process.on('message', resolve);
+      },
+    );
+    return address;
   }
 }
